Add data setup to download method

+++ vacation/data/dataloader.py
@@ -1,49 +1,106 @@
+import gc
+import shutil
 import sys
 import warnings
-from pathlib import Path
-
-import requests
-import shutil
 from hashlib import file_digest
+from pathlib import Path
 
 import h5py
 import numpy as np
+import requests
 import torch
 from torch.utils.data import Dataset
 
-def download_dataset(location: Path, overwrite: bool = True):
+
+def download_dataset(location: Path, redownload: bool = False, overwrite: bool = True):
 
     url = "https://zenodo.org/records/10845026/files/Galaxy10_DECals.h5"
-    
+
     if not isinstance(location, Path):
         location = Path(location)
 
     location.mkdir(exist_ok=True, parents=True)
     target = location / url.split("/")[-1]
 
-    if target.is_file():
-        if not overwrite:
-            warnings.warn("The data file already exists at this location! Use a different location or set 'overwrite=True'.")
-        else:
-            target.unlink(missing_ok=True)
-            print(f"Downloading data from {url}. Saving to {str(target)}.")
+    if target.is_file() and not redownload:
+        warnings.warn(
+            "The data file already exists at this location!"
+            "Use a different location or set 'redownload=True'."
+        )
+    else:
+        target.unlink(missing_ok=True)
+        target.touch()
+        print(f"Downloading data from {url}. Saving to {str(target)}.")
 
-            with requests.get(url, stream=True) as req:
-                with open(target, "wb") as file:
-                    shutil.copyfileobj(req.raw, file)
+        with requests.get(url, stream=True) as req:
+            with open(target, "wb") as file:
+                shutil.copyfileobj(req.raw, file)
 
-            print("Finished download of data!")
+        print("Finished download of data!")
 
     with open(target, "rb") as file:
         md5sum = file_digest(file, "md5").hexdigest()
 
     if md5sum != "c6b7b4db82b3a5d63d6a7e3e5249b51c":
         warnings.warn(
-            "The MD5 checksum of the downloaded file is not equal to the original file! " 
+            "The MD5 checksum of the downloaded file is not equal to the original file! "
             "Consider retrying the download!"
         )
     else:
         print("Checksum passed successfully.")
+
+    target_proc = Path(target.parent) / (target.stem + "_proc.h5")
+
+    if target_proc.is_file():
+        if overwrite:
+            target_proc.unlink()
+        elif not overwrite:
+            raise FileExistsError(
+                "This file already exists. Set 'overwrite = True' to overwrite the file!"
+            )
+
+    gc.collect()
+
+    with h5py.File(target, "r") as hf:
+
+        print("Reading image data ...")
+        images = np.asarray(hf["images"], dtype=np.float32)
+        print(
+            f"Completed reading image data (Memory size {images.nbytes * 1e-6} MB) ..."
+        )
+
+        images /= 255.0
+
+        print("Reading label data ...")
+        labels = np.asarray(hf["ans"], dtype=np.uint8)
+        print(
+            f"Completed reading label data (Memory size {labels.nbytes * 1e-6} MB) ..."
+        )
+
+    with h5py.File(target_proc, "w") as hf:
+        print("Writing image data ...")
+        hf.create_dataset(
+            "images",
+            data=images,
+            maxshape=(None, *images.shape[1:]),
+            chunks=True,
+            compression="gzip",
+        )
+
+        del images
+
+        print("Writing label data ...")
+        hf.create_dataset(
+            "ans",
+            data=labels,
+            maxshape=(None, *labels.shape[1:]),
+            chunks=True,
+            compression="gzip",
+        )
+
+        del labels
+
+    gc.collect()
 
 
 # Data Caching adapted from own previous project: simtools
